requirements.txt
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==5.1.1
exceptiongroup==1.1.1
executing==1.2.0
googlesearch-python==1.2.3
idna==3.4
iniconfig==2.0.0
ipython==8.13.2
jedi==0.18.2
lxml==4.9.2
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pytest==7.3.1
PyYAML==6.0
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
urllib3==2.0.2
wcwidth==0.2.6
